Move decorators into util file
