+++ clowder/clowder/cli/forall_controller.py
@@ -7,13 +7,11 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
+from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.cli.parallel import forall
 from clowder.cli.util import options_help_message
-from clowder.clowder_repo import (
-    print_clowder_repo_status,
-    valid_clowder_yaml_required
-)
-from clowder.cli.globals import CLOWDER_CONTROLLER
+from clowder.clowder_repo import print_clowder_repo_status
+from clowder.util.decorators import valid_clowder_yaml_required
 
 
 class ForallController(ArgparseController):
