+++ clowder/clowder/cli/prune_controller.py
@@ -12,10 +12,6 @@ import sys
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
-from clowder.clowder_repo import (
-    print_clowder_repo_status,
-    valid_clowder_yaml_required
-)
 from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.cli.util import (
     existing_branch_groups,
@@ -28,6 +24,8 @@ from clowder.cli.util import (
     validate_groups,
     validate_projects
 )
+from clowder.clowder_repo import print_clowder_repo_status
+from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.util.connectivity import network_connection_required
