+++ clowder/clowder/cli/reset_controller.py
@@ -7,13 +7,11 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.cli.parallel import reset
-from clowder.clowder_repo import (
-    print_clowder_repo_status_fetch,
-    valid_clowder_yaml_required
-)
 from clowder.cli.globals import CLOWDER_CONTROLLER
+from clowder.cli.parallel import reset
 from clowder.cli.util import options_help_message
+from clowder.clowder_repo import print_clowder_repo_status_fetch
+from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.util.connectivity import network_connection_required
