+++ clowder/clowder/cli/sync_controller.py
@@ -12,13 +12,11 @@ import sys
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
-from clowder.clowder_repo import (
-    print_clowder_repo_status_fetch,
-    valid_clowder_yaml_required
-)
 from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.cli.parallel import sync
 from clowder.cli.util import options_help_message
+from clowder.clowder_repo import print_clowder_repo_status_fetch
+from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.util.connectivity import network_connection_required
