+++ clowder/clowder/clowder_repo.py
@@ -11,10 +11,9 @@ import atexit
 import os
 import sys
 
-from termcolor import colored, cprint
+from termcolor import colored
 
 import clowder.util.formatting as fmt
-from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.error.clowder_error import ClowderError
 from clowder.error.clowder_yaml_error import ClowderYAMLError
 from clowder.git.project_repo import ProjectRepo
@@ -247,18 +246,6 @@ class ClowderRepo(object):
 CLOWDER_REPO = ClowderRepo(os.getcwd())
 
 
-def clowder_required(func):
-    """If no clowder repo, print clowder not found message and exit"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        _validate_clowder_repo_exists()
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
 def print_clowder_repo_status(func):
     """Print clowder repo status"""
 
@@ -281,31 +268,3 @@ def print_clowder_repo_status_fetch(func):
         return func(*args, **kwargs)
 
     return wrapper
-
-
-def valid_clowder_yaml_required(func):
-    """If clowder.yaml is invalid, print invalid yaml message and exit"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        _validate_clowder_repo_exists()
-        if CLOWDER_REPO.error:
-            print(fmt.invalid_yaml_error())
-            print(fmt.error(CLOWDER_REPO.error))
-            sys.exit(42)
-        if CLOWDER_CONTROLLER.error:
-            print(fmt.invalid_yaml_error())
-            print(fmt.error(CLOWDER_CONTROLLER.error))
-            sys.exit(42)
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def _validate_clowder_repo_exists():
-    """If clowder repo doesn't exist, print message and exit"""
-
-    if not os.path.isdir(CLOWDER_REPO.clowder_path):
-        cprint(' - No .clowder found in the current directory\n', 'red')
-        sys.exit(1)
